backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0
google-generativeai>=0.8.4
python-dotenv>=1.0.1
pydantic-settings>=2.0
pyjwt[crypto]>=2.8
cryptography>=42.0
httpx>=0.27

# Optional: export as Word (DOCX)
python-docx>=1.1.2
Pillow>=10.0.0

# Optional: export as PDF
reportlab>=4.0.0

# Optional: markdown to HTML (Doc Pro export)
markdown2>=2.4.0

# Optional: extract text from PDF (format-and-export pipeline)
pypdf>=4.0.0

# Optional: import videos from YouTube (server-side download)
yt-dlp>=2024.12.23

# Optional: Groq API (Llama 4 Vision + Whisper audio)
groq>=0.12.0
